Filter/Agg transpose bugfix (#252)
The check for whether or not a filter expression could be pushed beyond
an agg node was incorrect.

It was checking if the column was in the group by columns (checking the
equality of the numbers), when it should be checking based on indices if
we are only referring to columns that are *emitted* from the agg node as
group by columns.

For example, if we see:
```
Filter #1 > 100
    Agg { groups: [#1], agg: Sum() }
```

We should *not* push down because `#1` refers to the sum column. In the
current main branch, it is pushed down because it sees that `#1` equals
a column in the `groups` field. It should be checking that every column
is `< groups.len()` instead.