feat(core): add structured trace (#271)
(See memo_dump.planner.sql for the example)

Compared with the standard log-based trace, structured trace puts all
information related to one group together, therefore helping the
developer understand what's going on in the system. The trace is
appended during apply_rule and optimize_inputs with a step and a stage
number, indicating the order of the operations.

Signed-off-by: Alex Chi Z <iskyzh@gmail.com>